requirements.txt
click
awscli
flake8
black
python-dotenv>=0.20.0
pre-commit
jupyterlab
ipywidgets
tqdm
pandas
openpyxl
seaborn
spacy
nltk
googletrans==4.0.0rc1
opensearch-py
python-Levenshtein